room Managment

+++ HotelSystem/Models/Room.cs
@@ -6,7 +6,7 @@ namespace HotelSystem.Models
 	{
 		public string Name { get; set; }
 		public RoomType Type { get; set; }
-		public int PricePerNight { get; set; }
+		public decimal PricePerNight { get; set; }
 		public ICollection<RoomPicture> RoomPictures { get; set; }
 		public ICollection<RoomFacility> RoomFacilities { get; set; }
 		public ICollection<RoomOffer> RoomOffers { get; set; }
